requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
sentence-transformers>=2.2.0
pinecone-client>=3.0.0
feedparser>=6.0.0

# LLM Providers
anthropic>=0.18.0
openai>=1.10.0
google-genai>=0.3.1
groq>=0.4.0
ollama>=0.1.6
